Refactored models in fileupload to enable auto creation of license, cleanup
