+++ backend/core/admin.py
@@ -4,10 +4,10 @@ from django.utils import timezone
 from datetime import timedelta
 from django.contrib.admin import ModelAdmin
 
-from core.fileupload.models.family import Family
-from core.fileupload.models.license import License
-from core.fileupload.models.tag import Tag
-from core.fileupload.models.file import File
+from core.fileupload.models import Family
+from core.fileupload.models import License
+from core.fileupload.models import Tag
+from core.fileupload.models import File
 from core.user.forms import AdminUserChangeForm, AdminUserCreationForm
 from core.user.models import User
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
